Removed unneeded number widget, Moved InvalidConfigurationErrorMessageFactory.php
# Conflicts:
#	templates/bundles/PrestaShopBundle/Admin/TwigTemplateForm/prestashop_ui_kit_base.html.twig
