+++ src/PrestaShopBundle/Controller/Admin/Sell/Order/InvoicesController.php
@@ -139,7 +139,7 @@ class InvoicesController extends FrameworkBundleAdminController
             try {
                 $formHandler->save($form->getData());
             } catch (DataProviderException $e) {
-                $errorMessageFactory = $this->get('form.invalid_configuration_error_message_factory');
+                $errorMessageFactory = $this->get('form.error_message.invalid_configuration_error_message_factory');
                 $this->flashErrors($errorMessageFactory->getErrorMessages($e->getInvalidConfigurationDataErrors(), $form));
 
                 return false;
