+++ src/PrestaShopBundle/Form/ErrorMessage/InvalidConfigurationErrorMessageFactory.php
@@ -26,7 +26,7 @@
 
 declare(strict_types=1);
 
-namespace PrestaShopBundle\Form;
+namespace PrestaShopBundle\Form\ErrorMessage;
 
 use PrestaShopBundle\Controller\Exception\FieldNotFoundException;
 use PrestaShopBundle\Entity\Repository\LangRepository;
