+++ src/PrestaShopBundle/Resources/config/services/bundle/form/form.yml
@@ -5,11 +5,5 @@ services:
     form.form_cloner:
         class: PrestaShopBundle\Form\FormCloner
 
-    form.invalid_configuration_error_message_factory:
-        class: PrestaShopBundle\Form\InvalidConfigurationErrorMessageFactory
-        arguments:
-           - '@translator'
-           - '@prestashop.core.admin.lang.repository'
-
     form.form_builder_modifier:
         class: PrestaShopBundle\Form\FormBuilderModifier
