+++ templates/bundles/PrestaShopBundle/Admin/TwigTemplateForm/prestashop_ui_kit_base.html.twig
@@ -1208,10 +1208,3 @@
   {{- block('form_widget_simple') -}}
   <img src="{{ form.vars.value }}" alt="{{ ('Image preview for ' ~ form.vars.name)|trim }}" />
 {% endblock %}
-
-{%- block number_widget -%}
-  {# type="number" doesn't work with floats #}
-  {%- set type = type|default('text') -%}
-  {{ block('form_widget_simple') }}
-  {{ block('form_help') }}
-{%- endblock number_widget -%}
